Remove ununsed MeetingTimeRemainingUpdateEvtMsg and BreakoutRoomsTimeRemainingUpdateEvtMsg
