+++ akka-bbb-apps/src/main/scala/org/bigbluebutton/core/apps/breakout/UpdateBreakoutRoomsTimeMsgHdlr.scala
@@ -1,7 +1,7 @@
 package org.bigbluebutton.core.apps.breakout
 
 import org.bigbluebutton.common2.msgs._
-import org.bigbluebutton.core.api.{ SendTimeRemainingAuditInternalMsg, UpdateBreakoutRoomTimeInternalMsg }
+import org.bigbluebutton.core.api.UpdateBreakoutRoomTimeInternalMsg
 import org.bigbluebutton.core.apps.{ PermissionCheck, RightsManagementTrait }
 import org.bigbluebutton.core.bus.BigBlueButtonEvent
 import org.bigbluebutton.core.db.{ BreakoutRoomDAO, MeetingDAO, NotificationDAO }
@@ -87,9 +87,6 @@ trait UpdateBreakoutRoomsTimeMsgHdlr extends RightsManagementTrait {
 
       val event = buildUpdateBreakoutRoomsTimeEvtMsg(msg.body.timeInMinutes)
       outGW.send(event)
-
-      //Force Update time remaining in the clients
-      eventBus.publish(BigBlueButtonEvent(props.meetingProp.intId, SendTimeRemainingAuditInternalMsg(props.meetingProp.intId, msg.body.timeInMinutes)))
 
       updatedModel match {
         case Some(model) => {
