+++ akka-bbb-apps/src/main/scala/org/bigbluebutton/core/running/MeetingActorAudit.scala
@@ -73,9 +73,6 @@ class MeetingActorAudit(
     eventBus.publish(BigBlueButtonEvent(props.meetingProp.intId, MonitorNumberOfUsersInternalMsg(props.meetingProp.intId)))
     eventBus.publish(BigBlueButtonEvent(props.meetingProp.intId, MonitorGuestWaitPresenceInternalMsg(props.meetingProp.intId)))
 
-    // Trigger updating users of time remaining on meeting.
-    eventBus.publish(BigBlueButtonEvent(props.meetingProp.intId, SendTimeRemainingAuditInternalMsg(props.meetingProp.intId, 0)))
-
     if (props.meetingProp.isBreakout) {
       // This is a breakout room. Update the main meeting with list of users in this breakout room.
       eventBus.publish(BigBlueButtonEvent(
